feat(network): moved constants.ts
