+++ src/app/networkDiscovery/constants.ts
