+++ src/app/networkDiscovery/views/index.ts
@@ -1,0 +1,2 @@
+export { default as DiscoveriesList } from "./DiscoveriesList";
+export { default as NetworkDiscoveryConfigurationForm } from "./NetworkDiscoveryConfigurationForm";
